fix #209 : created a viewer package and splited module viewer.py into view.py, model.py and api.py
combo.py also moved in viewer package
