+++ larray/__init__.py
@@ -6,15 +6,4 @@ from larray.ufuncs import *
 from larray.excel import open_excel
 from larray.ipfp import ipfp
 from larray.example import load_example_data
-
-try:
-    from larray.viewer import view, edit, compare
-except ImportError:
-    def view(*args, **kwargs):
-        raise Exception('view() is not available because Qt is not installed')
-
-    def edit(*args, **kwargs):
-        raise Exception('edit() is not available because Qt is not installed')
-
-    def compare(*args, **kwargs):
-        raise Exception('compare() is not available because Qt is not installed')
+from larray.viewer import *
